rename env to env.server
